5697: Implement EventStartEvent notification pub/sub & refactor
- Remove /builder package (bad naming and contains classes AbstractEventListener & MessageBuilder that were already implemented)
