+++ src/main/java/faang/school/notificationservice/config/context/RedisConfig.java
@@ -1,6 +1,6 @@
 package faang.school.notificationservice.config.context;
 
-import faang.school.notificationservice.listener.ProfileViewEventListener;
+import faang.school.notificationservice.messaging.ProfileViewEventListener;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
