+++ src/main/java/faang/school/notificationservice/messaging/AbstractEventListener.java
@@ -3,7 +3,7 @@ package faang.school.notificationservice.messaging;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import faang.school.notificationservice.client.UserServiceClient;
 import faang.school.notificationservice.dto.UserDto;
-import faang.school.notificationservice.service.MessageBuilder;
+import faang.school.notificationservice.service.message_builder.MessageBuilder;
 import faang.school.notificationservice.service.NotificationService;
 import lombok.RequiredArgsConstructor;
 
@@ -11,19 +11,19 @@ import java.util.List;
 import java.util.Locale;
 
 @RequiredArgsConstructor
-public abstract class AbstractEventListener {
+public abstract class AbstractEventListener<T> {
 
-    private final ObjectMapper objectMapper;
-    private final UserServiceClient userServiceClient;
-    private final List<MessageBuilder> messageBuilders;
+    protected final ObjectMapper objectMapper;
+    protected final UserServiceClient userServiceClient;
+    private final List<MessageBuilder<T>> messageBuilders;
     private final List<NotificationService> notificationServices;
 
-    protected String getMessage(Class<?> eventType, Locale userLocale, String... args) {
+    protected String getMessage(T event, Locale userLocale) {
         return messageBuilders.stream()
-                .filter(messageBuilder -> messageBuilder.supportsEventType(eventType))
+                .filter(messageBuilder -> messageBuilder.supportsEventType(event))
                 .findFirst()
-                .map(messageBuilder -> messageBuilder.buildMessage(eventType, userLocale,args))
-                .orElseThrow(() -> new IllegalArgumentException("No message builder found for the given event type: " + eventType.getName()));
+                .map(messageBuilder -> messageBuilder.buildMessage(event, userLocale))
+                .orElseThrow(() -> new IllegalArgumentException("No message builder found for the given event type: " + event.getClass().getName()));
     }
 
     protected void sendNotification(Long id, String message) {
