+++ src/main/java/faang/school/notificationservice/service/message_builder/MessageBuilder.java
@@ -1,8 +1,9 @@
-package faang.school.notificationservice.builder;
+package faang.school.notificationservice.service.message_builder;
 
 import java.util.Locale;
 
 public interface MessageBuilder<T> {
+
     String buildMessage(T eventType, Locale locale);
     boolean supportsEventType(T eventType);
 }
