+++ src/main/java/faang/school/notificationservice/service/message_builder/ProfileViewMessageBuilder.java
@@ -1,6 +1,7 @@
-package faang.school.notificationservice.builder;
+package faang.school.notificationservice.service.message_builder;
 
 import faang.school.notificationservice.dto.ProfileViewEvent;
+import faang.school.notificationservice.service.message_builder.MessageBuilder;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.MessageSource;
